Automate Xcode Index/Processing
Script will now open Xcode and wait a period for it to index and
process the project so that we can run xcodebuild. The wait time
is in seconds and can be set with indexWaitTime. The script then
closes Xcode and continues the normal build.

This had to be added because I found that newly generated cordova
projects were missing some required project files after initially
being generated, and opening the project in Xcode is the only way
I have found to fix that problem.