Update script and support files for XCode 8.3.3
Since XCode 8.3.3 the PackageApplication utility is no longer shipped
witht he product. There are work arounds to this, but they aren't
sustainable in the long run, and we needed to switch over to the latest
supported approach using `exportArchive`.

This adds example build.json and export-ipa.plist files to support the
changes required in how new xcode projects are signed and archived in
the recent versions.

The README is updated to walk through finding and setting the placeholder
values for TEAM ID and provisioning profile name. This step needs to be
automated, and split out into another file, but I'm only one person, doing
this as I find time and energy.